refactor : input컴포넌트 로직은 최대한 빼고 ui만 남겨놓기
