+++ components/moduleComponents/signIn/signin.module.css
@@ -10,3 +10,15 @@
   position: absolute;
   bottom: 10px;
 }
+
+
+.eye_icon{
+  position: absolute;
+  right: 15px;
+  bottom: 36px;
+}
+
+.error_Input{
+  color: red;
+  border-color: red;
+}
