+++ next.config.js
@@ -2,7 +2,12 @@
 const nextConfig = {
   reactStrictMode: true,
   images: {
-    domains:['ca.slack-edge.com','codeit-images.codeit.com'],
+    remotePatterns:[
+      {
+        protocol: 'https',
+        hostname: '*'
+      }
+    ]
   }
 };
