+++ pages/folder.tsx
@@ -21,3 +21,5 @@ export default function FolderPage() {
     </>
   );
 }
+
+//다이나믹 라우팅으로 만들어야할듯
