+++ pages/signin.tsx
@@ -10,3 +10,4 @@ const Singin = () => {
 };
 
 export default Singin;
+
